Fix stringcommon.h

+++ Chowdren/base/stringcommon.h
@@ -59,7 +59,7 @@ inline std::string number_to_string(int value)
     return fast_itoa(value);
 }
 
-inline std::string number_to_string(size_t value)
+inline std::string number_to_string(unsigned int value)
 {
     return fast_itoa(value);
 }
